requirements.txt
django>=4.2.0
djangorestframework>=3.16.0
discord.py>=2.3.0
python-decouple>=3.8
django-environ>=0.11.2
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
whitenoise>=6.6.0
django-cors-headers>=4.3.1
django-crispy-forms>=2.1
crispy-bootstrap5>=2023.10
gunicorn>=21.2.0
cryptography>=42.0.7